chore: refactor hooks that export useContext call, update types, remove barrel
