+++ frontend/hooks/useSettings.ts
@@ -2,7 +2,4 @@ import { useContext } from 'react';
 
 import { SettingsContext } from '@/context/SettingsProvider';
 
-export const useSettings = () => {
-  const { screen, goto } = useContext(SettingsContext);
-  return { screen, goto };
-};
+export const useSettings = () => useContext(SettingsContext);
