+++ frontend/hooks/useStore.ts
@@ -2,8 +2,4 @@ import { useContext } from 'react';
 
 import { StoreContext } from '@/context/StoreProvider';
 
-export const useStore = () => {
-  const { storeState } = useContext(StoreContext);
-
-  return { storeState };
-};
+export const useStore = () => useContext(StoreContext);
