+++ frontend/hooks/useWallet.ts
@@ -2,14 +2,4 @@ import { useContext } from 'react';
 
 import { WalletContext } from '@/context/WalletProvider';
 
-export const useWallet = () => {
-  const { wallets, masterEoaAddress, masterSafeAddress, updateWallets } =
-    useContext(WalletContext);
-
-  return {
-    wallets,
-    masterEoaAddress,
-    masterSafeAddress,
-    updateWallets,
-  };
-};
+export const useWallet = () => useContext(WalletContext);
